+++ tests/test_nova_api/test_cases.py
@@ -3,9 +3,10 @@ import unittest
 import os
 import uuid
 from datetime import datetime
+import time
 
 from itk_dev_shared_components.kmd_nova.authentication import NovaAccess
-from itk_dev_shared_components.kmd_nova.nova_objects import NovaCase, CaseParty
+from itk_dev_shared_components.kmd_nova.nova_objects import NovaCase, CaseParty, Caseworker, Department
 from itk_dev_shared_components.kmd_nova import nova_cases
 
 
@@ -57,6 +58,18 @@ class NovaCasesTest(unittest.TestCase):
             name="Test Test"
         )
 
+        caseworker = Caseworker(
+            name='svcitkopeno svcitkopeno',
+            ident='AZX0080',
+            uuid='0bacdddd-5c61-4676-9a61-b01a18cec1d5'
+        )
+
+        department = Department(
+            id=818485,
+            name="Borgerservice",
+            user_key="4BBORGER"
+        )
+
         case = NovaCase(
             uuid=str(uuid.uuid4()),
             title=f"Test {datetime.now()}",
@@ -65,10 +78,34 @@ class NovaCasesTest(unittest.TestCase):
             case_parties=[party],
             kle_number="23.05.01",
             proceeding_facet="G01",
-            sensitivity="Fortrolige"
+            sensitivity="Fortrolige",
+            caseworker=caseworker,
+            responsible_department=department,
+            security_unit=department
         )
 
         nova_cases.add_case(case, self.nova_access)
+
+        # Wait up to 10 seconds for the case to be created in Nova
+        for _ in range(10):
+            time.sleep(1)
+            cases = nova_cases.get_cases(self.nova_access, cpr=party.identification, case_title=case.title)
+            if cases:
+                break
+
+        self.assertEqual(len(cases), 1)
+
+        nova_case = cases[0]
+        self.assertEqual(nova_case.uuid, case.uuid)
+        self.assertEqual(nova_case.title, case.title)
+        self.assertEqual(nova_case.progress_state, case.progress_state)
+        self.assertEqual(nova_case.kle_number, case.kle_number)
+        self.assertEqual(nova_case.proceeding_facet, case.proceeding_facet)
+        self.assertEqual(nova_case.sensitivity, case.sensitivity)
+        self.assertEqual(nova_case.caseworker.ident, case.caseworker.ident)
+        self.assertEqual(nova_case.case_parties[0].identification, case.case_parties[0].identification)
+        self.assertEqual(nova_case.responsible_department.id, case.responsible_department.id)
+        self.assertEqual(nova_case.security_unit.id, case.security_unit.id)
 
 
 if __name__ == '__main__':